Make the symbol table a HashMap instead of a SlotMap
* While the SlotMap might seem efficient at first, the usecase here
makes it less so. Whenever a Variable is encountered when evaluating an
expression, the value for the variable has to be looked up by something.
Right now, I don't think there is a way to parse a string and associate
it with the key from a slot map. I'd have to somehow store the key and
the string and find the key and overall have roundabout ways of
accessing values